requirements.txt
# ============================================
# Dependencias - HotelApp API
# ============================================

# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de Datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Validación
pydantic==2.5.0

# Plantillas y Forms
jinja2==3.1.2
python-multipart==0.0.6

# Seguridad (para futuras implementaciones)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Variables de Entorno
python-dotenv==1.0.0

# CORS
python-dateutil==2.8.2